# jaxrs-spec-generated/swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "OiOi content management API",
    "description" : "OiOi content management API spec",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/v1"
  } ],
  "security" : [ {
    "BearerAuth" : [ ]
  } ],
  "paths" : {
    "/devices" : {
      "get" : {
        "tags" : [ "Devices" ],
        "summary" : "List devices",
        "description" : "List devices",
        "operationId" : "listDevices",
        "responses" : {
          "200" : {
            "description" : "Devices",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Device"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Devices" ],
        "summary" : "Create a device",
        "description" : "Create device",
        "operationId" : "createDevice",
        "requestBody" : {
          "description" : "Payload",
          "content" : {
            "application/json;charset=utf-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/Device"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Created device",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Device"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/devices/{deviceId}" : {
      "get" : {
        "tags" : [ "Devices" ],
        "summary" : "Finds a device",
        "description" : "Finds a device",
        "operationId" : "findDevice",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "device id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Found device",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Device"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Devices" ],
        "summary" : "Updates device",
        "description" : "Updates a device",
        "operationId" : "updateDevice",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "device id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "description" : "Payload",
          "content" : {
            "application/json;charset=utf-8" : {
              "schema" : {
                "$ref" : "#/components/schemas/Device"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Updated device",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Device"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Devices" ],
        "summary" : "Delete device",
        "description" : "Delete a device",
        "operationId" : "deleteDevice",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "device id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Successful removal"
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Device" : {
        "required" : [ "name" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Device id.",
            "format" : "uuid",
            "readOnly" : true
          },
          "name" : {
            "type" : "string",
            "description" : "Device name."
          },
          "createdAt" : {
            "type" : "string",
            "description" : "Creation time",
            "format" : "date-time",
            "readOnly" : true
          },
          "modifiedAt" : {
            "type" : "string",
            "description" : "Last modification time",
            "format" : "date-time",
            "readOnly" : true
          }
        }
      },
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "description" : "The error message",
            "readOnly" : true
          }
        }
      }
    },
    "securitySchemes" : {
      "BearerAuth" : {
        "type" : "http",
        "scheme" : "bearer"
      }
    }
  }
}